Tea rewrite.
DO NOT USE THIS IN PRODUCTION!

Rewriting Tea to work better and be more organized. All of the command parsing logic will be done in the tea executable instead of the individual classes, simplifing classes and making tea DRYer.
User class was rewritten to work with new tea executable. Other classes will run into issues (won't work or won't work as planned).

Also redid gitignore as it will be staying and won't be deleted on tea init.

Changed include_once to include on TFD\Core\Router so routes will always be loaded.